requirements.txt
# Python version: 3.12.9
gymnasium==1.1.1
numpy==2.2.4
PyYAML==6.0.2
pandas==2.2.3
scipy==1.15.2
tqdm==4.67.1
